requirements.txt
streamlit
crewai
crewai-tools
PyPDF2
python-dotenv
groq
litellm
pandas